fix(S3Audit): Add a check for misconfigured S3 bucket CNAME
With CNAME'd buckets, the hostname and the bucket name must be the
same, however if the CNAME points to a different bucket which does
exist, the previous version of the auditor would not catch it correctly.

This update adds an additional check to ensure that the CNAME'd value
and the name of the bucket / hostname is the same.